+++ packages/vue/src/pagination/pagination-context.ts
@@ -1,9 +1,7 @@
-import { type connect } from '@zag-js/pagination'
-import { type ComputedRef } from 'vue'
 import { createContext } from '../context'
 import { type UsePaginationReturn } from './use-pagination'
 
 export const [PaginationProvider, usePaginationContext] =
-  createContext<ComputedRef<ReturnType<typeof connect>>>('PaginationContext')
+  createContext<UsePaginationReturn>('PaginationContext')
 
 export type PaginationContext = UsePaginationReturn